==> 1_datapreparation.ipynb <==
#imports
import csv
import requests
import json
import os      # for files
import csv     # csv file handling
import codecs  # because text file encoding is an issue in 2018
from datetime import datetime, date #time

competitionID = 119 # AT Bundesliga
seasonID = 2018     #2018/2019
requestURLMatchStatic = F"http://omo.akamai.opta.net/competition.php?competition={competitionID}&season_id={seasonID}&feed_type=F1&user=Hackweekend&psw=35BJ10nwf&json"
r = requests.get(requestURLMatchStatic
                         ,timeout=None
                         ,headers={  'Content-Type': 'application/json'}
                        )    
jsondata = r.json()
docMatch = jsondata['SoccerFeed']['SoccerDocument']['MatchData']
#get all matches
matchList = []
for match in docMatch:
    matchObj = match["@attributes"]
#    print(matchObj)
    matchID_long = match["@attributes"]["uID"]
    matchID = matchID_long.replace('g','')
    matchList.append(matchID)
print(matchList)

#get all matches with winners
u=1
allrows=[]
print("trying to import matches: "+str(len(matchList)))
while u<len(matchList):    
    gameID=matchList[u]
    print("Parsing GameID: "+str(gameID))
    u=u+1
    r = requests.get(F"http://omo.akamai.opta.net/?game_id={gameID}&feed_type=F9&user=Hackweekend&psw=35BJ10nwf&json"
                         ,timeout=None
                         ,headers={  'Content-Type': 'application/json'}
                        )    
    jsondata = r.json()
    if 'SoccerFeed' in jsondata:
        matchData = jsondata['SoccerFeed']['SoccerDocument']['MatchData']
        teamdata = matchData['TeamData']
        onerow = {}
        onerow['matchid']=gameID;
        onerow['matchdata']=matchData['MatchInfo']['Date']
        potwinner=matchData['MatchInfo']['Result']['@attributes']
        if 'Winner' not in potwinner:
            onerowwinner='none'
        else:
            onerowwinner = matchData['MatchInfo']['Result']['@attributes']['Winner']
        teamRefs = []
        #statreflist for one time setting
        statreflist= [];
        i=0
        while i<len(teamdata):
            teamstats = teamdata[i] 
            onerow['team_'+str(i)+'_teamRef']= teamstats['@attributes']['TeamRef']
            teamstatslist = teamstats['Stat']
            teamRefs.append(teamstats['@attributes']['TeamRef'])
            if len(statreflist)==0:
                l=0
                while l<len(teamstatslist):
                    stat = teamstatslist[l]
                    stattype = stat['@attributes']['Type']
                    #TODO what up with fh and sh separation?            
                    statvalue = stat['@value']
                    onerow['team_'+str(i)+'_s_'+stattype] = float(statvalue)
                    statreflist.append(stattype);
                    l=l+1
            else :
                usedstats=[]
                l=0
                while l<len(teamstatslist):
                    stat = teamstatslist[l]
                    stattype = stat['@attributes']['Type']
                    statvalue = stat['@value']
                    if stattype in statreflist:
                        usedstats.append(stattype)
                        onerow['team_'+str(i)+'_s_'+stattype] = float(statvalue)
                    l=l+1
                #anything that is in statreflist and not in usedstats needs value 0   
                openstats = list(set(statreflist).difference(set(usedstats)))
                if len(openstats)>0:
                    k=0;
                    while k<len(openstats):
                        onerow['team_'+str(i)+'_s_'+openstats[k]] = 0
                        k=k+1
            i=i+1
        #who is winner team_0 or team_1
        z=0
        while z<len(teamRefs):
            if teamRefs[z]==onerowwinner:
                onerow['winner']='team_'+str(z)
            z=z+1
        if 'winner' not in onerow:
            onerow['winner']='draw'
        allrows.append(onerow)

print("FOUND: "+str(len(allrows)))

import pandas as pd
allMatchData = pd.DataFrame.from_dict(allrows)
allMatchData.fillna(0, inplace=True)

allMatchData.head()

list(allMatchData)

allMatchData.describe()

allMatchData.dtypes

# Parse all Events with matchid

## Coding Table preparation

import os 
cwd = os.getcwd()
print(cwd)
import_file = "./../db/qualifiers.csv" 
#ID;Qualifier;Values;Description
qualifiers_dict = {}
with codecs.open(import_file, 'rU','cp1252') as handle:
    reader = csv.DictReader(handle, delimiter=";", quotechar='"')
    for line in reader:
        qualifiers_dict[line["ID"]] = {'Qualifier' : line["Qualifier"], 'Values' : line['Values'],'Description' : line['Description']}
        
import_file = "./../db/events.csv" 
#Event ID;Name;Description
events_dict = {}
with codecs.open(import_file, 'rU','cp1252') as handle:
    reader = csv.DictReader(handle, delimiter=";", quotechar='"')
    for line in reader:
        events_dict[line["Event ID"]] = {'Name' : line['Name'], 'Description' : line['Description']}
        
#print(qualifiers_dict)
#print(qualifiers_dict['7']) # Players caught offside
#print(events_dict['50'])    # Dispossessed

competitionID = 119 # AT Bundesliga
seasonID = 2018     #2018/2019
requestURLSquad = F"http://omo.akamai.opta.net/competition.php?competition={competitionID}&season_id={seasonID}&feed_type=F40&user=Hackweekend&psw=35BJ10nwf&json"
r = requests.get(requestURLSquad
                         ,timeout=None
                         ,headers={  'Content-Type': 'application/json'}
                        )    
jsondata = r.json()
docSquad = jsondata['SoccerFeed']['SoccerDocument']['Team']
matchSquad = docSquad[1]
#print(json.dumps(matchSquad, indent=4, sort_keys=True))
playerDict = {}
for team in docSquad:
    teamObj = team["@attributes"]
    teamName = teamObj['short_club_name']
    teamID = teamObj['uID'].replace('t','')
    for player in team["Player"]:
        playerID_long = player["@attributes"]["uID"]
        playerID = playerID_long.replace('p','')
        playerName = player["Name"]
        playerDict[playerID] = {'Name' : playerName, 'Team' : teamName, 'TeamID' : teamID}
#print(playerDict)

# Get All Events for all Matches in Our Match DataFram

parsedMatches = list(allMatchData['matchid'])

interested in successfull events per eventtype per team 

so we need to know for the event the team and if it is a success (if it is team_0 or team 1

gameid='988751'
refRow = allMatchData[allMatchData['matchid'] == gameid]
refValues = refRow[['matchid','winner','team_1_teamRef','team_0_teamRef']]

refValues

i = 0
events=[]
while i<len(parsedMatches):
    gameID = parsedMatches[i]
    print("parsing Game: "+str(gameID))
    requestURL = F"http://omo.akamai.opta.net/?game_id={gameID}&feed_type=F24&user=Hackweekend&psw=35BJ10nwf&json"
    r = requests.get(requestURL
                         ,timeout=None
                         ,headers={  'Content-Type': 'application/json'}
                        )    
    jsondata = r.json()
    game = jsondata['Games']['Game']    
    gamedata = game['@attributes']['game_date']
    doc = game['Event']
    #get teamRef and stats for this match
    refRow = allMatchData[allMatchData['matchid'] == gameID]
    refValues = refRow[['matchid','winner','team_1_teamRef','team_0_teamRef']]
    #print(refValues)
    #now we know the teamrefs it concerns the part
    #what do we want for each event?
    #per each eventtype the sum of all events before that where sucessfull for 
    #each team?
    #so we make a ref table - which has:
    #team -> team0 or team1 - teamid , eventtype, successfull    
    for event in doc:
    ## direct attributes
        eventline = {}
        eventAttr = event['@attributes']
        outcome = int(eventAttr['outcome'])
        eventline['matchid'] = gameID
        eventline['timestamp'] = eventAttr['timestamp']
        eventline['id'] = eventAttr['id']
        eventline['outcome'] = int(eventAttr['outcome'])
        eventline['type'] = eventAttr['type_id']
        eventline['winner'] = refValues['winner'].values[0]
        # player from ID (if involved)
        playerInvolved = False
        playerName = ""
        try:
            playerID = eventAttr['player_id']
            playerObj = playerDict[playerID]
            playerName = playerObj['Name']
            teamID = "t"+playerObj['TeamID']
            teamName = playerObj['Team']
            playerInvolved = True
        except KeyError:
            playerName = 'NO PLAYER'
            playerInvolved = False
        if playerInvolved:
            team = ''
            isTeam1 = refValues[refValues['team_1_teamRef']==teamID];
            if len(isTeam1)>0:
                team='team_1'
            else: 
                team='team_0'
            eventline['team'] = team
            eventline['teamid'] = teamID
        else : 
            eventline['team'] = 'NONE'
            eventline['teamid'] = 'NONE'
        events.append(eventline)
    i=i+1

allEventsData = pd.DataFrame.from_dict(events)

# Now We have all Events data 
- we need for each event the sum for each eventtype
- for this team 
-> where outcome = 1 and matchid = matchid

len(list(allEventsData.type.unique()))

allEventsData['datetest'] = pd.to_datetime(allEventsData['timestamp'])

allEventsData.head()

len(allEventsData)

allEventsData.to_csv("allEventsData.csv", encoding="utf-8") 

allMatchData.to_csv("allMatchData.csv", encoding="utf-8")
