==> 1_dataprepartion_eventssum.ipynb <==
#load files needed
import pandas as pd

allEventsData=pd.read_csv("allEventsData.csv", encoding="utf-8", index_col=0) 
allEventsData['datetest'] = pd.to_datetime(allEventsData['timestamp'])

allEventsData.head()

allEventsData.dtypes

allMatchData = pd.read_csv("allMatchData.csv", encoding="utf-8", index_col=0) 

allMatchData.head()

parsedMatches = list(allMatchData['matchid'])

len(parsedMatches)

Reforamt Events for ml

eventtypelist = list(allEventsData.type.unique())
print("eventtypes: "+str(len(eventtypelist)))
parsedMatches = list(allMatchData['matchid'])
print("parsedMatches: "+str(len(parsedMatches)))
mlrows = []
gameIndex = 0;
for gameID in parsedMatches:
    print(str(gameIndex)+"start transfrom match: "+str(gameID))
    matchData = allEventsData[allEventsData['matchid']==gameID]
    eventIdList = list(matchData['id'])
    print(str(gameIndex)+"transforming events :"+str(len(eventIdList)))
    eventIndex = 0
    for eventId in eventIdList:
        #get all events before this event  
        if eventIndex%100==0:
            print(str(gameIndex)+" done event "+str(eventIndex))
        row = matchData[(matchData['id']==eventId)]
        currenttimestamp = row['datetest'].values[0]
        toadd = {}
        toadd['team']=row['team'].values[0]    
        toadd['type']=row['type'].values[0]
        toadd['timestamp']=row['timestamp'].values[0]
        toadd['winner']=row['winner'].values[0]
        toadd['teamid']=row['datetest'].values[0]
        toadd['matchid']=gameID
        toadd['id']=eventId
        #for each types for each team the sum
        previousData = matchData[(matchData['matchid']==gameID) \
                                             & (matchData['datetest']<currenttimestamp) \
                                             & (matchData['outcome']==1)]
        for eventtype in eventtypelist:
            #make it for team 1 for testing
            previousRows = previousData[(previousData['team']=='team_0') \
                                            & (previousData['type']==eventtype)]
            toadd['team_0_'+str(eventtype)] = len(previousRows)
            previousRows = previousData[(previousData['team']=='team_1') \
                                            & (previousData['type']==eventtype)]
            toadd['team_1_'+str(eventtype)] = len(previousRows)
        mlrows.append(toadd)
        eventIndex=eventIndex+1
    print(str(gameIndex)+"end transfrom match: "+str(gameID))
    gameIndex = gameIndex+1

eventsMlData = pd.DataFrame.from_dict(mlrows)

eventsMlData.to_csv("eventsMlData.csv", encoding="utf-8") 

eventsMlData['teamid']=allEventsData['teamid']

eventsMlData['outcome']=allEventsData['outcome']

eventsMlData['timestamp']=allEventsData['timestamp']

eventsMlData.head()

eventsMlData.to_csv("eventsMlData.csv", encoding="utf-8") 

len(eventsMlData)
