==> 2_datapreparation_mlpreviousmatch.ipynb <==
#load files needed
import pandas as pd

allEventsData=pd.read_csv("allEventsData.csv", encoding="utf-8", index_col=0) 
allEventsData['datetest'] = pd.to_datetime(allEventsData['timestamp'])

allEventsData.head()

print(allEventsData.dtypes)
print(len(allEventsData))

allMatchData = pd.read_csv("allMatchData.csv", encoding="utf-8", index_col=0) 
allMatchData['matchdate'] = pd.to_datetime(allMatchData['matchdata'])

allMatchData.head()

What are good indicators that describe the match?

- team_0|1_games_won_before
- stats of game before in same position?

mlMatchData = allMatchData[['matchid','matchdata','team_0_teamRef','team_1_teamRef','winner']].copy()
mlMatchData['matchdate'] = pd.to_datetime(mlMatchData['matchdata'])

mlMatchData.head()

mlMatchData.dtypes

def games_won_before(row,df,team):
    matchdate = row['matchdate']
    teamRef = row[team+'_teamRef']
    win = len(df[ ( df['matchdate']< matchdate ) & \
        ( df['winner']==team) & (df[team+'_teamRef']==teamRef) ])  
    #print(win)
    return win

mlMatchData['winner'].unique()

mlMatchData['team_0_won_before'] = mlMatchData.apply(lambda x: games_won_before(x,mlMatchData,'team_0'), axis=1)
mlMatchData['team_1_won_before'] = mlMatchData.apply(lambda x: games_won_before(x,mlMatchData,'team_1'), axis=1)

mlMatchData.head()

test = mlMatchData.mean(axis=0)

fields = list(allMatchData.columns[2:-1])
finalfields = [ x.replace("team_0","") for x in fields if "team_1" not in x ]
finalfields =finalfields[:-1]
len(finalfields)

#how many passes has this team made in the game before this game ( single)
def getLastWinMatchStat(row,df,curfield,team):
    teamRef = row[team+'_teamRef']
    matchdate = row['matchdate']
    lastWonMatch=df[ ( df['matchdate']< matchdate ) & \
        ( df['winner']==team) & (df[team+'_teamRef']==teamRef) ].sort_values(by=['matchdate'],ascending=False)
    if len(lastWonMatch)>0:
        return lastWonMatch.mean()[team+curfield]
    else:
        return 0

for curfield in finalfields:
    mlMatchData['team_0_lastwinmatch_'+curfield] = mlMatchData.apply(lambda x: getLastWinMatchStat(x,allMatchData,curfield,'team_0'), axis=1)
    mlMatchData['team_1__lastwinmatch_'+curfield] = mlMatchData.apply(lambda x: getLastWinMatchStat(x,allMatchData,curfield,'team_1'), axis=1)

mlMatchData.to_csv("mlMatchData.csv", encoding="utf-8") 

len(mlMatchData)
