==> 3_ml.ipynb <==
import pandas as pd

allMatchData = pd.read_csv("allMatchData.csv", encoding="utf-8", index_col=0) 
allMatchData['matchdate'] = pd.to_datetime(allMatchData['matchdata'])

allMatchData.head()

allMatchData.shape

eventsMlData=pd.read_csv("eventsMlData.csv", encoding="utf-8", index_col=0) 

eventsMlData.head()

eventsMlData.shape

mlMatchData= pd.read_csv("mlMatchData.csv", encoding="utf-8", index_col=0) 

mlMatchData.head()

mlMatchData.shape

- join eventsmldata  with ml Matchdata

fields = list(allMatchData.columns[2:-2])
finalfields = [ x.replace("team_0","") for x in fields if "team_1" not in x ]
finalfields =finalfields[:-1]
len(finalfields)

mergeaddfieldnames = ['matchid','team_0_won_before','team_1_won_before'];
for curfield in finalfields:
    mergeaddfieldnames.append('team_0_lastwinmatch_'+curfield)
    mergeaddfieldnames.append('team_1__lastwinmatch_'+curfield)    

len(mergeaddfieldnames)

sourceDf = pd.merge(eventsMlData,mlMatchData[mergeaddfieldnames],on='matchid', how='left')

sourceDf.dtypes

#remove fields
removefields = ['id','matchid','matchid','timestamp','team','winner']
topredict='winner'
mlInput = sourceDf.drop(removefields, axis=1)

mlInput.head()

from sklearn.feature_extraction import DictVectorizer
v = DictVectorizer(sparse=False)
X = v.fit_transform(mlInput.to_dict(orient='records'))
from sklearn.preprocessing import Normalizer
transformer = Normalizer().fit(X)
Y=transformer.transform(X)

from sklearn import preprocessing
classesToTransfrom= list(sourceDf[topredict])
le = preprocessing.LabelEncoder()
classes = le.fit_transform(classesToTransfrom)

X.shape

classes.shape

from sklearn.linear_model import LogisticRegression
clf = LogisticRegression(max_iter=400,random_state=0, solver='lbfgs',multi_class='ovr').fit(X, classes)

probs=clf.predict_proba(X)

classLabels =le.inverse_transform(clf.classes_)

#add prediction back to source dataset
dfpred = pd.DataFrame(probs, columns=classLabels)

for label in classLabels:
    dfpred[label] = dfpred[label].apply(lambda x: round(x, 4)*100)
    sourceDf['win_perc_'+label]=dfpred[label]

sourceDf.head()

sourceDf[ ['id','timestamp','matchid', 'team', 'type', 'winner', 'win_perc_draw','win_perc_team_0','win_perc_team_1'] ]

sourceDf.to_csv("predictions.csv", encoding="utf-8")
